requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0

# Weaviate (подключение к вашему VDS)
weaviate-client>=4.0.0

# Для оценки RAGAS (опционально)
datasets>=2.14.0
ragas>=0.2.0
pydantic>=2.0.0,<3.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
